services/auth-service/requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.34.0
sqlalchemy==2.0.49
asyncpg==0.31.0
alembic==1.18.4
bcrypt==5.0.0
python-jose==3.5.0
pydantic==2.12.5
